+++ CHANGELOG.md
@@ -37,6 +37,8 @@ A preview of the next release can be installed from
 
 </details>
 
+- [#2151](https://github.com/babashka/babashka/issues/2151): An `:exec-fn` task in the `:depends` of a task with a `:task` body runs, with its own `:exec-args` and spec defaults
+
 ## 1.13.222 (2026-09-14)
 
 Read the release blog post [here](https://blog.michielborkent.nl/babashka-1.13.222.html).